// shipeasy/src/app/models/container-master.ts
 export interface Container {
   imoDate: string;
    eta: string;
    booking: string;
  imoNo: string;
    igmDate: string;
    terminalIn: string;
    reject: string;
    shippingBillNumber: string;
    netWeight: any;
    mastercontainerId: string;
    imoType: string;
    icdOut: string;
    icdIn: string;
    grossWeight: string;
    factoryOut: any;
    factoryIn: string;
    evgmNumber: string;
    doDate: string;
    depotOut: string;
    bondNumber: string;
    batchId: string;
    cargoType: string;
    dischargeDate: string;
    sobDate: string;
    igmNumber: string;
    blDate: string;
    containerNumber: string;
    blNumber: string;
    batchNo: string;
    voyageNo: string;
    containerId: string;
    vesselName: string;
    shippingLineName: string;
    name: string;
    partymasterId: string;
    systemtypeId: string;
    typeName: string;
    containerNo: string;
    cargoNo: string;
    cargoTypeName: string;
    containerTypeId: string;
    containerType: string;
    containerSize: string;
    tankStatusId: string;
    tankStatus: string;
    tankType: string;
    tarWeight: string;
    tankCapacity: string;
    exitOffHireDate: string;
    onHireDate: string;
    dateOfManufacture: string;
    oneWay: boolean;
    loadCapacity: string;
    containerOperator: string;
    yard: string;
    yardName: string;
    pickLocation: string;
    dropLocation: string;
    soc: boolean;
    maxGrossWeight: string;
    maxPayload: string;
    baffles: boolean;
    remarks: string;
    status: boolean;
    module: string;
    referenceId: string;
    containermasterId: string;
    createdOn: string;
    updatedOn: string;
    tenantId: string;
    createdBy: string;
    createdByUID: string;
    updatedBy: string;
    updatedByUID: string;
    customerName:string;
    containerStatus:string;
    previousStatus:string

  }

export  interface Location {
 
    locationName: string;
    portType: string;
    country: string;
    state: string;
    masterType: string;
    agentBranch: string;
    CFS: boolean;
    ICD: boolean;
    Yard: boolean;
    name: string;
    code: string;
    portName: string;
    terminal: string;
    EDICode: string;
    DOaddress: string;
    address: string;
    contactPerson: string;
    email: string;
    primaryCountryCode: string;
    primaryNo: string;
    DOCode: string;
    bondNo: string;
    creditDays: string;
    lineReference: boolean;
    countryISOCode: string;
    stateId: string;
    portId: string;
    status: boolean;
    module: string;
    tenantId: string;
    agentBranchId: string;
    referenceId: string;
    locationId: string;
    createdOn: string;
    updatedOn: string;
    createdBy: string;
    createdByUID: string;
    updatedBy: string;
    updatedByUID: string;
  }
  
  